Using unsigned short instead of unsigned.

+++ instruction.h
@@ -8,7 +8,7 @@ class Instruction
 {
 public:
     virtual ~Instruction() = 0;
-    virtual void operator()(std::vector<float>& memory, unsigned& subprogram_pointer, std::vector<unsigned>& instruction_pointers, unsigned& return_pointer) const = 0;
+    virtual void operator()(std::vector<float>& memory, unsigned short& subprogram_pointer, std::vector<unsigned short>& instruction_pointers, unsigned short& return_pointer) const = 0;
 protected:
     static float clamp(float value);
 };
