+++ instructions/increase.cpp
@@ -1,12 +1,12 @@
 #include "increase.h"
 
-Increase::Increase(unsigned input_pointer, unsigned output_pointer) :
+Increase::Increase(unsigned short input_pointer, unsigned short output_pointer) :
     m_input_pointer(input_pointer),
     m_output_pointer(output_pointer)
 {
 }
 
-void Increase::operator()(std::vector<float>& memory, unsigned& subprogram_pointer, std::vector<unsigned>& instruction_pointers, unsigned& return_pointer) const
+void Increase::operator()(std::vector<float>& memory, unsigned short& subprogram_pointer, std::vector<unsigned short>& instruction_pointers, unsigned short& return_pointer) const
 {
     (void)return_pointer;
     memory[m_output_pointer] = memory[m_input_pointer] + 1;
