+++ instructions/loop.h
@@ -6,10 +6,10 @@
 class Loop : public Instruction
 {
 public:
-    Loop(unsigned subprogram_pointer);
-    void operator()(std::vector<float>& memory, unsigned& subprogram_pointer, std::vector<unsigned>& instruction_pointers, unsigned& return_pointer) const override;
+    Loop(unsigned short subprogram_pointer);
+    void operator()(std::vector<float>& memory, unsigned short& subprogram_pointer, std::vector<unsigned short>& instruction_pointers, unsigned short& return_pointer) const override;
 private:
-    unsigned m_subprogram_pointer;
+    unsigned short m_subprogram_pointer;
 };
 
 #endif // LOOP_H
