+++ instructions/not.cpp
@@ -1,12 +1,12 @@
 #include "not.h"
 
-Not::Not(unsigned input_pointer, unsigned output_pointer) :
+Not::Not(unsigned short input_pointer, unsigned short output_pointer) :
     m_input_pointer(input_pointer),
     m_output_pointer(output_pointer)
 {
 }
 
-void Not::operator()(std::vector<float>& memory, unsigned& subprogram_pointer, std::vector<unsigned>& instruction_pointers, unsigned& return_pointer) const
+void Not::operator()(std::vector<float>& memory, unsigned short& subprogram_pointer, std::vector<unsigned short>& instruction_pointers, unsigned short& return_pointer) const
 {
     (void)return_pointer;
     memory[m_output_pointer] = !memory[m_input_pointer];
