+++ instructions/return.h
@@ -6,7 +6,7 @@
 class Return : public Instruction
 {
 public:
-    void operator()(std::vector<float>& memory, unsigned& subprogram_pointer, std::vector<unsigned>& instruction_pointers, unsigned& return_pointer) const override;
+    void operator()(std::vector<float>& memory, unsigned short& subprogram_pointer, std::vector<unsigned short>& instruction_pointers, unsigned short& return_pointer) const override;
 };
 
 #endif // RETURN_H
